@@ -1,7 +1,8 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>net5.0</TargetFramework>
+    <TargetFramework>net6.0</TargetFramework>
+    <Configurations>Debug;Release;ReleaseX86;ReleaseX64;Release-ServiceX86;Release-ServiceX64</Configurations>
   </PropertyGroup>
 
 </Project>
