@@ -11,19 +11,19 @@ namespace Agent.Commands
     public class DeleteDirectoryCommand : AgentCommand
     {
         public override string Name => "rmdir";
-        public override void InnerExecute(AgentTask task, Models.Agent agent, AgentTaskResult result, CommModule commm)
+        public override void InnerExecute(AgentTask task, AgentCommandContext context)
         {
             string path;
             if (task.SplittedArgs.Length != 1)
             {
-                result.Result = $"Usage : {this.Name} folder_to_delete";
+                context.Result.Result = $"Usage : {this.Name} folder_to_delete";
                 return;
             }
                 
             path = task.SplittedArgs[0];
 
             var dirInfo = Directory.CreateDirectory(path);
-            result.Result = $"{dirInfo.FullName} created";
+            context.Result.Result = $"{dirInfo.FullName} created";
         }
     }
 }
