@@ -12,25 +12,26 @@ namespace Agent.Commands
     {
         public override string Name => "download";
 
-        public override void InnerExecute(AgentTask task, Models.Agent agent, AgentTaskResult result, CommModule commm)
+        public override void InnerExecute(AgentTask task, AgentCommandContext context)
         {
-            var fileName = task.FileName;
+            this.CheckFileDownloaded(task, context);
 
-            var fileContent = commm.Download(task.FileId, a =>
+            var file = context.FileService.ConsumeDownloadedFile(task.FileId);
+            string fileName = file.Name;
+
+            if (task.SplittedArgs.Count() > 0)
             {
-                result.Info = $"Downloading {fileName} ({a}%)";
-                commm.SendResult(result);
-            }).Result;
+                fileName = task.SplittedArgs[0];
+            }
 
-
-            this.Notify(result, commm, $"{fileName} Downloaded");
+            var fileContent = file.GetFileContent();
 
             using (FileStream fs = new FileStream(fileName, FileMode.Create, FileAccess.Write))
             {
                 fs.Write(fileContent, 0, fileContent.Length);
             }
 
-            result.Result = $"File dowloaded to {fileName}.";
+            context.Result.Result = $"File downloaded to {fileName}." + Environment.NewLine;
         }
     }
 }
