@@ -10,9 +10,9 @@ namespace Agent.Commands
     public class TerminateCommand: AgentCommand
     {
         public override string Name => "terminate";
-        public override void InnerExecute(AgentTask task, Models.Agent agent, AgentTaskResult result, CommModule commm)
+        public override void InnerExecute(AgentTask task, AgentCommandContext context)
         {
-            result.Result = "Exiting...";
+            context.Result.Result = "Exiting...";
             Environment.Exit(0);
         }
     }
