@@ -12,10 +12,11 @@ namespace Agent.Commands
     {
         public override string Name => "whoami";
 
-        public override void InnerExecute(AgentTask task, Models.Agent agent, AgentTaskResult result, CommModule commm)
+        public override void InnerExecute(AgentTask task, AgentCommandContext context)
         {
             var identity = WindowsIdentity.GetCurrent();
-            result.Result = identity.Name;
+            context.AppendResult(identity.Name);
+            context.Error("Whoami failed!");
         }
     }
 }
