@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Runtime.Serialization;
 using System.Text;
 using System.Threading.Tasks;
 
@@ -14,7 +15,11 @@ namespace Agent.Models
         public string ProcessName { get; set; }
         public int ProcessId { get; set; }
         public string Integrity { get; set; }
-
         public string Architecture { get; set; }
+        public string EndPoint { get; set; }
+        public string Version { get; set; }
+
+        public int SleepInterval { get; set; }
+        public int SleepJitter { get; set; }
     }
 }
