@@ -27,6 +27,8 @@ namespace Commander.Commands.Agent
         public override string Name => "interact";
         public override ExecutorMode AvaliableIn => ExecutorMode.All;
 
+        public override string[] Alternate { get => new string[1] { "int" }; }
+
         public override RootCommand Command => new RootCommand(this.Description)
             {
                 new Argument<string>("id", "index or id of the agent"),
@@ -39,7 +41,7 @@ namespace Commander.Commands.Agent
             if (int.TryParse(context.Options.id, out index))
                 agent = context.CommModule.GetAgent(index);
             else
-                agent = context.CommModule.GetAgents().FirstOrDefault(a => a.Metadata.ShortId.ToLower().Equals(context.Options.id.ToLower()));
+                agent = context.CommModule.GetAgents().FirstOrDefault(a => a.Metadata.Id.ToLower().Equals(context.Options.id.ToLower()));
             
             if(agent == null)
             {
@@ -49,8 +51,9 @@ namespace Commander.Commands.Agent
 
             context.Executor.CurrentAgent = agent;
             context.Executor.Mode = ExecutorMode.AgentInteraction;
+            var star = agent.Metadata.Integrity == "High" ? "*" : string.Empty;
 
-            context.Terminal.Prompt = $"${ExecutorMode.Agent} {agent.Metadata.UserName}@{agent.Metadata.Hostname}> ";
+            context.Terminal.Prompt = $"${ExecutorMode.Agent}({agent.Metadata.Id}) {agent.Metadata.UserName}{star}@{agent.Metadata.Hostname}> ";
 
             return true;
         }
