@@ -1,6 +1,8 @@
 ﻿using Commander.Communication;
 using Commander.Executor;
+using Commander.Models;
 using Commander.Terminal;
+using Spectre.Console;
 using System;
 using System.Collections.Generic;
 using System.CommandLine;
@@ -27,33 +29,37 @@ namespace Commander.Commands.Agent
         protected async override Task<bool> HandleCommand(CommandContext<StatusCommandOptions> context)
         {
             var agent = context.Executor.CurrentAgent;
-            if (agent.LastSeen.AddSeconds(30) > DateTime.UtcNow)
-                context.Terminal.WriteSuccess($"Agent {agent.Metadata.ShortId} is up and running !");
+            if (agent.IsActive == true)
+                context.Terminal.WriteSuccess($"Agent {agent.Metadata.Id} is up and running !");
             else
-                context.Terminal.WriteError($"Agent {agent.Metadata.ShortId} seems to be not responding!");
-            var results = new SharpSploitResultList<StatusResult>();
-            results.Add(new StatusResult() { Name = "Id", Value = agent.Metadata.ShortId });
-            results.Add(new StatusResult() { Name = "Hostname", Value = agent.Metadata.Hostname });
-            results.Add(new StatusResult() { Name = "UserName", Value = agent.Metadata.UserName });
-            results.Add(new StatusResult() { Name = "ProcessId", Value = agent.Metadata.ProcessId.ToString() });
-            results.Add(new StatusResult() { Name = "ProcessName", Value = agent.Metadata.ProcessName });
-            results.Add(new StatusResult() { Name = "Architecture", Value = agent.Metadata.Architecture });
-            results.Add(new StatusResult() { Name = "Integrity", Value = agent.Metadata.Integrity });
-            results.Add(new StatusResult() { Name = "Last Seen", Value = agent.LastSeen.ToLocalTime().ToString("dd/MM/yyyy hh:mm:ss") });
-            context.Terminal.WriteLine(results.ToString());
+                context.Terminal.WriteError($"Agent {agent.Metadata.Id} seems to be not responding!");
+
+            var table = new Table();
+            table.Border(TableBorder.Rounded);
+            // Add some columns
+            table.AddColumn(new TableColumn("Item").LeftAligned());
+            table.AddColumn(new TableColumn("Value").LeftAligned());
+            table.HideHeaders();
+
+            table.AddRow("Id", agent.Metadata?.Id ?? string.Empty);
+            table.AddRow("Hostname", agent.Metadata?.Hostname ?? string.Empty);
+            table.AddRow("User Name", agent.Metadata?.UserName ?? string.Empty);
+            table.AddRow("Process Id", agent.Metadata?.ProcessId.ToString());
+            table.AddRow("Process Name", agent.Metadata?.ProcessName ?? string.Empty);
+            table.AddRow("Architecture", agent.Metadata?.Architecture ?? string.Empty);
+            table.AddRow("Integrity", agent.Metadata?.Integrity ?? string.Empty);
+            table.AddRow("EndPoint", agent.Metadata?.EndPoint ?? string.Empty);
+            table.AddRow("Version", agent.Metadata?.Version ?? string.Empty);
+            table.AddRow("Sleep", agent.Metadata?.Sleep ?? string.Empty);
+            table.AddRow("First Seen", agent.FirstSeen.ToLocalTime().ToString());
+            table.AddRow("Last Seen", Math.Round(agent.LastSeenDelta.TotalSeconds,2).ToString() + "s" ?? string.Empty);
+            
+
+            context.Terminal.Write(table);
             return true;
         }
-    }
 
-    public sealed class StatusResult : SharpSploitResult
-    {
-        public string Name { get; set; }
-        public string Value { get; set; }
 
-        protected internal override IList<SharpSploitResultProperty> ResultProperties => new List<SharpSploitResultProperty>()
-            {
-                new SharpSploitResultProperty { Name = nameof(Name), Value = Name },
-                new SharpSploitResultProperty { Name = nameof(Value), Value = Value },
-            };
+        
     }
 }
