@@ -3,6 +3,8 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using Spectre.Console;
+using Spectre.Console.Rendering;
 
 namespace Commander.Terminal
 {
@@ -13,6 +15,8 @@ namespace Commander.Terminal
         bool CanHandleInput { get; set; }
 
         string Prompt { get; set; }
+
+        //public ConsoleColor DefaultColor { get; set; }
         Task Start();
 
         void stop();
@@ -21,12 +25,12 @@ namespace Commander.Terminal
 
         void Restore();
 
-        void NewLine(bool lineBreak = true);
+        void NewLine(bool brk = true);
 
 
         void WriteLine(TerminalMessageType typ, params string[] strs);
 
-        void WritePrompt();
+       // void WritePrompt();
 
         public void WriteSuccess(params string[] parm);
         public void WriteError(params string[] parm);
@@ -35,8 +39,15 @@ namespace Commander.Terminal
         void WriteLine();
 
         void Write(string text);
+        void Write(string text, int count);
 
         void WriteLine(params string[] parm);
+
+        void Write(IRenderable renderabel);
+
+        void WriteMarkup(string markup);
+
+        void WriteLineMarkup(string markup);
 
         void SaveCursorPosition();
 
