@@ -7,12 +7,14 @@ using System.Threading.Tasks;
 
 namespace Commander.Terminal
 {
+
+
+
     public partial class Terminal : ITerminal
     {
         public const string DefaultPrompt = "$> ";
 
         public event EventHandler<string> InputValidated;
-        public ConsoleColor DefaultColor { get; set; } = Console.ForegroundColor;
 
         CancellationTokenSource _token = new CancellationTokenSource();
 
@@ -24,6 +26,8 @@ namespace Commander.Terminal
         }
         public async Task Start()
         {
+            //this.WriteLine(Console.WindowWidth + "-" + Console.WindowHeight);
+            this.NewLine(false);
             while (!_token.IsCancellationRequested)
             {
                 if (Console.KeyAvailable)
@@ -43,216 +47,135 @@ namespace Commander.Terminal
                     }
                 }
 
-                await Task.Delay(10);
+                await Task.Delay(2);
             }
         }
 
-        private int _positionInLine = 0;
-        private string _currentLine = string.Empty;
-
-        private List<string> _history = new List<string>();
-        private int _historyPosition = 0;
-
-       
+        private CommandHistory History = new CommandHistory();
 
         public string Prompt { get; set; } = "$> ";
 
-        protected int CursorLeft
-        {
-            get
-            {
-                return this.Prompt.Length + this._positionInLine;
-            }
-        }
+        private CommandDetail CurrentCommand { get; set; }
+
         protected void HandleKey(ConsoleKeyInfo key)
         {
+           
+
             if (!this.CanHandleInput)
                 return;
             switch (key.Key)
             {
-                case ConsoleKey.LeftArrow:
+                case ConsoleKey.LeftArrow: this.CurrentCommand.HandleInput(CommandDetail.HandledKey.LeftArrow); break;
+                case ConsoleKey.RightArrow: this.CurrentCommand.HandleInput(CommandDetail.HandledKey.RightArrow); break;
+                case ConsoleKey.Home: this.CurrentCommand.HandleInput(CommandDetail.HandledKey.Home); break;
+                case ConsoleKey.End: this.CurrentCommand.HandleInput(CommandDetail.HandledKey.End); break;
+                case ConsoleKey.Backspace: this.CurrentCommand.HandleInput(CommandDetail.HandledKey.BackSpace); break;
+                case ConsoleKey.Delete: this.CurrentCommand.HandleInput(CommandDetail.HandledKey.Delete); break;
+                default:
                     {
-                        if (_positionInLine > 0)
+                        if (key.Key == ConsoleKey.C && (key.Modifiers & ConsoleModifiers.Control) != 0)
                         {
-                            Console.CursorLeft -= 1;
-                            _positionInLine -= 1;
+                            this.History.Pop();
+                            this.NewLine();
+                            break;
                         }
-                    }
-                    break;
-                case ConsoleKey.RightArrow:
-                    {
-                        if (_positionInLine < this._currentLine.Length)
-                        {
-                            Console.CursorLeft += 1;
-                            _positionInLine += 1;
-                        }
-                    }
-                    break;
-                case ConsoleKey.Home:
-                    {
-                        _positionInLine = 0;
-                        Console.CursorLeft = this.CursorLeft;
-                    }
-                    break;
-                case ConsoleKey.End:
-                    {
-                        _positionInLine = this._currentLine.Length;
-                        Console.CursorLeft = this.CursorLeft;
-                    }
-                    break;
 
+                        this.CurrentCommand.HandleInput(key.KeyChar);
+                    }
+                    break;
                 case ConsoleKey.UpArrow:
                     {
-                        if (_historyPosition < _history.Count -1)
-                        {
-                            _historyPosition++;
-                            LoadHistory();
-                        }
+                        var cmd = this.History.Previous();
+                        if (cmd != null)
+                            this.CreateNewCommandAndPrint(true, cmd.Value);
                     }
                     break;
                 case ConsoleKey.DownArrow:
                     {
-                        if (_historyPosition > 0)
-                        {
-                            _historyPosition--;
-                            LoadHistory();
-                        }
+                        var cmd = this.History.Next();
+                        if (cmd != null)
+                            if (this.History.IsMostRecent(cmd))
+                            {
+                                this.CurrentCommand.Interrupt();
+                                cmd.CursorStartY = this.CurrentCommand.CursorStartY;
+                                this.CurrentCommand = cmd;
+                                this.CurrentCommand.Print();
+                            }
+                            else
+                                this.CreateNewCommandAndPrint(true, cmd.Value);
                     }
                     break;
                 case ConsoleKey.Enter:
                     {
                         //Save to history
-                        string line = this._currentLine.Trim();
-                        Console.WriteLine();
+                        var cmd = this.CurrentCommand;
+
+                        string line = this.CurrentCommand.Value.Trim();
+
                         if (!string.IsNullOrEmpty(line))
                         {
-                            this.RefreshCurrentHistory();
+                            this.History.Pop();
+                            this.History.Register(cmd);
+                            Console.WriteLine();
                             this.InputValidated?.Invoke(this, line);
                         }
                         else
                         {
-                            this._history.RemoveAt(0);
-                            this.NewLine(false);
-                        }
-                    }
-                    break;
-                case ConsoleKey.Backspace:
-                    {
-                        if (_positionInLine > 0)
-                        {
-                            Console.CursorLeft -= 1;
-                            _positionInLine -= 1;
-                            this._currentLine = this._currentLine.Remove(this._positionInLine, 1);
-                            Console.CursorLeft = this.CursorLeft;
-                            Console.Write(LineRightPart + " ");
-                            Console.CursorLeft = this.CursorLeft;
-                            this.RefreshCurrentHistory();
-                        }
-                    }
-                    break;
-                case ConsoleKey.Delete:
-                    {
-                        if (_positionInLine < this._currentLine.Length)
-                        {
-                            this._currentLine = this._currentLine.Remove(this._positionInLine, 1);
-                            Console.Write(LineRightPart + " ");
-                            Console.CursorLeft = this.CursorLeft;
-                            this._history.RemoveAt(0);
-                            this.RefreshCurrentHistory();
-                        }
-                    }
-                    break;
-                default:
-                    {
-                        if (key.Key == ConsoleKey.C && (key.Modifiers & ConsoleModifiers.Control) != 0)
-                        {
-                            this._history.RemoveAt(0);
+                            this.History.Pop();
                             this.NewLine();
-                            break;
                         }
-
-
-                        var c = key.KeyChar;
-                        string str = string.Empty;
-                        str += c;
-
-                        this._currentLine = this._currentLine.Insert(this._positionInLine, str);
-                        WriteAndResetCursor(this.LineRightPart);
-                        _positionInLine += 1;
-                        Console.CursorLeft = this.CursorLeft;
-
-                        this.RefreshCurrentHistory();
                     }
                     break;
-
-
             }
         }
 
-        protected string LineRightPart
+
+        //protected void WriteAndResetCursor(string str)
+        //{
+        //    Console.Write(str);
+        //    Console.CursorLeft = this.CursorLeft;
+        //}
+
+        private void CreateNewCommandAndPrint(bool replace = false, string cmd = null)
         {
-            get
+            int top = Console.CursorTop;
+            if (replace)
             {
-                return this._currentLine.Substring(this._positionInLine, this._currentLine.Length - this._positionInLine);
+                this.CurrentCommand.Interrupt();
+                top = this.CurrentCommand.CursorStartY;
             }
+
+            this.CurrentCommand = new CommandDetail(top, this.Prompt, cmd);
+            this.CurrentCommand.Print();
         }
 
-        protected void WriteAndResetCursor(string str)
+        public void NewLine(bool brk = true)
         {
-            Console.Write(str);
-            Console.CursorLeft = this.CursorLeft;
-        }
-
-        public void NewLine(bool lineBreak = true)
-        {
-            if (lineBreak)
+            if (brk)
                 Console.WriteLine();
-            this.WritePrompt();
-            _positionInLine = 0;
-            _currentLine = string.Empty;
-            this._history.Insert(0, this._currentLine);
-            this._historyPosition = 0;
+            this.CreateNewCommandAndPrint();
+            this.History.Register(this.CurrentCommand);
         }
+
         public void stop()
         {
             _token.Cancel();
         }
 
-        private void LoadHistory()
-        {
-            var history = _history[_historyPosition];
-            int max = Math.Max(history.Length, this._currentLine.Length);
-            this._currentLine = history;
-            this._positionInLine = 0;
-            Console.CursorLeft = CursorLeft;
-            Console.Write(this._currentLine.PadRight(max));
-            this._positionInLine = this._currentLine.Length;
-            Console.CursorLeft = CursorLeft;
-        }
-
-        private void RefreshCurrentHistory()
-        {
-            if (this._history.Count > 0)
-                this._history.RemoveAt(0);
-            this._history.Insert(0, this._currentLine);
-        }
 
         public void Interrupt()
         {
-            Console.CursorLeft = 0;
-            Console.Write(string.Empty.PadRight(this.Prompt.Length + this._currentLine.Length));
-            Console.CursorLeft = 0;
+            this.CurrentCommand.Interrupt();
         }
 
         public void Restore()
         {
-            this.WritePrompt();
-            Console.Write(this._currentLine);
-            Console.CursorLeft = this.CursorLeft;
+            this.CurrentCommand.Reset(Console.CursorTop);
+            this.CurrentCommand.Print();
         }
 
-       
 
-        
+
+
     }
 }
