@@ -7,12 +7,11 @@ namespace TeamServer
 {
     public static class Logger
     {
-        public static bool Active { get; set; } //= true;
+        public static bool Active { get; set; } = true;
 
         public static string FileName { get; set; } = "log.log";
         public static void Log(string message)
         {
-            var Active = true;
             if (!Active)
                 return;
 
