@@ -3,13 +3,23 @@ using System.Collections.Concurrent;
 using System.Collections.Generic;
 using System.Linq;
 
+
 namespace TeamServer.Models
 {
     public class Agent
     {
-        public AgentMetadata Metadata { get; }
+
+
+
+        public string Id { get; set; }
+        public string RelayId { get; set; }
+
+        public List<string> Path { get; set; } = new List<string>();
+        public AgentMetadata Metadata { get; set; }
 
         public DateTime LastSeen { get; protected set; }
+
+        public DateTime FirstSeen { get; protected set; }
 
         public string ListenerId { get; set; }
 
@@ -18,10 +28,100 @@ namespace TeamServer.Models
         private readonly List<AgentTaskResult> _taskResults = new();
         public ConcurrentBag<AgentTask> TaskHistory { get; private set; } = new();
 
-        public Agent(AgentMetadata metadata)
+        private ConcurrentDictionary<string, ConcurrentQueue<SocksMessage>> _InboudSocksMessages = new();
+        private ConcurrentQueue<SocksMessage>_OutboundSocksMessages = new();
+
+
+        public Queue<AgentFileChunck> FileChuncksToUpload { get; private set; } = new Queue<AgentFileChunck>();
+
+        public Queue<SocksMessage> GetProxyResponses(string destination)
         {
-            this.Metadata = metadata;
+            var q = new Queue<SocksMessage>();
+            if (!_InboudSocksMessages.ContainsKey(destination))
+                return q;
+
+            while (_InboudSocksMessages[destination].TryDequeue(out var message))
+                q.Enqueue(message);
+
+            return q;
         }
+
+        public void AddProxyResponses(IEnumerable<SocksMessage> messages)
+        {
+            if (messages == null || !messages.Any())
+                return;
+
+            var src = messages.First().Source;
+            ConcurrentQueue<SocksMessage> q = null;
+
+            if (!_InboudSocksMessages.ContainsKey(src))
+            {
+                q = new ConcurrentQueue<SocksMessage>();
+                _InboudSocksMessages.TryAdd(src, q);
+            }
+            else
+                q = _InboudSocksMessages[src];
+
+            foreach(var m in messages)
+                q.Enqueue(m);
+        }
+
+
+        public List<SocksMessage> GetProxyRequests()
+        {
+            var q = new List<SocksMessage>();
+            while (_OutboundSocksMessages.TryDequeue(out var message))
+                q.Add(message);
+            return q;
+        }
+
+        public void SendProxyRequest(SocksMessage message)
+        {
+            _OutboundSocksMessages.Enqueue(message);
+        }
+
+        private AgentFileChunck GetNextChunck()
+        {
+            if (!FileChuncksToUpload.Any())
+                return null;
+            return FileChuncksToUpload.Dequeue();
+        }
+
+        public MessageTask GetNextMessage()
+        {
+            var mess = new MessageTask();
+            mess.Header.Owner = this.Id;
+            mess.Items.AddRange(this.GetPendingTaks());
+            mess.FileChunk = this.GetNextChunck();
+
+            mess.ProxyMessages = this.GetProxyRequests();
+
+            if (mess.FileChunk == null && !mess.Items.Any() && !mess.ProxyMessages.Any())
+                return null;
+
+            return mess;
+        }
+
+        public Agent(string id)
+        {
+            this.Metadata = new AgentMetadata();
+            this.Metadata.Id = id;
+            this.Id = id;
+            this.FirstSeen = DateTime.Now;
+        }
+
+        public void QueueDownload(List<AgentFileChunck> chunks)
+        {
+            foreach (var chunk in chunks)
+                this.FileChuncksToUpload.Enqueue(chunk);
+        }
+
+
+
+        //public Agent(AgentMetadata metadata)
+        //{
+        //    this.Metadata = metadata;
+        //}
 
         public void CheckIn()
         {
