@@ -17,8 +17,6 @@ namespace TeamServer.Models
 
         public virtual int BindPort { get; protected set; }
 
-        public virtual int PublicPort { get; protected set; }
-
         public virtual string Protocol { get; protected set; }
 
         protected IAgentService _agentService;
@@ -26,29 +24,34 @@ namespace TeamServer.Models
         protected IBinMakerService _binMakerService;
         protected IListenerService _listenerService;
         protected ILogger _logger;
+        protected IChangeTrackingService _changeTrackingService;
+        protected IWebHostService _webHostService;
+        protected ICryptoService _cryptoService;
+        protected IAuditService _auditService;
 
-        public Listener(string name, int bindPort, string Ip, int? publicPort = null)
+        public Listener(string name, int bindPort, string Ip)
         {
             this.Name = name;
             this.Ip = Ip;
             this.BindPort = bindPort;
-            if (!publicPort.HasValue)
-                this.PublicPort = bindPort;
-            else
-                this.PublicPort = publicPort.Value;
 
             this.Id = Guid.NewGuid().ToString();
         }
 
-        public abstract string Uri { get; }
-
-        public void Init(IAgentService service, IFileService fileService, IBinMakerService binMakerService, IListenerService listenerService, ILogger logger)
+        public void Init(IAgentService service, IFileService fileService, IBinMakerService binMakerService, IListenerService listenerService, ILogger logger, IChangeTrackingService changeTrackingService,
+            IWebHostService webHostService,
+            ICryptoService cryptoService,
+            IAuditService auditService)
         {
             this._agentService = service;
             this._fileService = fileService;
             this._binMakerService = binMakerService;
             this._listenerService = listenerService;
             this._logger = logger;
+            this._changeTrackingService = changeTrackingService;
+            this._webHostService = webHostService;
+            this._cryptoService = cryptoService;
+            this._auditService = auditService;
         }
 
         public abstract Task Start();
