@@ -6,8 +6,17 @@ using TeamServer.Models;
 
 namespace TeamServer.Services
 {
+    public interface IListenerService
+    {
+        void AddListener(Listener listener);
+        IEnumerable<Listener> GetListeners();
+        Listener GetListener(string id);
+        void RemoveListener(Listener listener);
+    }
+
     public class ListenerService : IListenerService
     {
+
         private readonly List<Listener> _listeners = new List<Listener>();
 
         public void AddListener(Listener listener)
@@ -15,9 +24,9 @@ namespace TeamServer.Services
             _listeners.Add(listener);
         }
 
-        public Listener GetListener(string name)
+        public Listener GetListener(string id)
         {
-            return this.GetListeners().FirstOrDefault(l => l.Name.Equals(name, StringComparison.OrdinalIgnoreCase));
+            return this.GetListeners().FirstOrDefault(l => l.Id.Equals(id, StringComparison.OrdinalIgnoreCase));
         }
 
         public IEnumerable<Listener> GetListeners()
